Add Erb processing of source before conversion
Process source files with Erb before parsing, to enable different
variations without needing to branch/edit.

Also add support fot "features", which can be enabled using config or
with a command line switch. Erb processing is done in the context of an
object that responds to any method (via method_missing) with true if
the corresponding feature has been enabled, false otherwise. This allows
for use such as:

    <% if my_feature %>
    This only includeed if 'my_feature' is enabled.
    <% end %>

and can be enabled on the command line with:

    cvgen -e my_feature infile

or setting the "features" option in the config (it should be an array).

Features can be enabled from the options or the command line and will be
combined if set in both.