Add "sections" support
Add create_section and section methods to Erb context to allow capturing
sections of text that can be included in different positions in the
document. create_section takes a name and a block, the block simply
being the markup you want to capture. It then stores this to be
retrieved later with the section method. The section method takes a name
and returns the previously captured chunk with that name.